Factor generic social messaging functions out into Social
