+++ Sources/App/Core/Twitter.swift
@@ -14,22 +14,14 @@
 
 import Fluent
 import OhhAuth
-import SemanticVersion
 import Vapor
 
 
 enum Twitter {
 
     private static let apiUrl: String = "https://api.twitter.com/1.1/statuses/update.json"
-    private static let tweetMaxLength = 260  // exactly 280 is rejected, plus leave some room for unicode accounting oddities
+    static let tweetMaxLength = 260  // exactly 280 is rejected, plus leave some room for unicode accounting oddities
     
-    enum Error: LocalizedError {
-        case invalidMessage
-        case missingCredentials
-        case postingDisabled
-        case requestFailed(HTTPStatus, String)
-    }
-
     struct Credentials {
         var apiKey: (key: String, secret: String)
         var accessToken: (key: String, secret: String)
@@ -37,7 +29,7 @@ enum Twitter {
 
     static func post(client: Client, tweet: String) async throws {
         guard let credentials = Current.twitterCredentials() else {
-            throw Error.missingCredentials
+            throw Social.Error.missingCredentials
         }
         let url: URL = URL(string: "\(apiUrl)?status=\(tweet.urlEncodedString())")!
         let signature = OhhAuth.calculateSignature(
@@ -53,115 +45,7 @@ enum Twitter {
         headers.add(name: "Content-Type", value: "application/x-www-form-urlencoded")
         let response = try await client.post(URI(string: url.absoluteString), headers: headers)
         guard response.status == .ok else {
-            throw Error.requestFailed(response.status, response.body?.asString() ?? "")
-        }
-    }
-
-}
-
-
-// MARK:- Helpers to post package to firehose
-
-extension Twitter {
-
-    static func createMessage(preamble: String,
-                              separator: String = "–",
-                              summary: String? = nil,
-                              url: String = "") -> String {
-        let link = "\n\n\(url)"
-        let separator = " \(separator) "
-        let availableLength = tweetMaxLength - preamble.count - separator.count - link.count
-        let description: String = {
-            guard let summary = summary?.trimmingCharacters(in: .whitespacesAndNewlines),
-                  !summary.isEmpty else { return "" }
-            let ellipsis = "…"
-            return summary.count < availableLength
-                ? separator + summary
-                : separator + String(summary.prefix(availableLength - ellipsis.count)) + ellipsis
-        }()
-
-        return preamble + description + link
-    }
-
-    static func newPackageMessage(packageName: String,
-                                  repositoryOwnerName: String,
-                                  url: String,
-                                  summary: String?) -> String {
-        createMessage(preamble: "📦 \(repositoryOwnerName) just added a new package, \(packageName)",
-                      summary: summary,
-                      url: url)
-    }
-
-    static func versionUpdateMessage(packageName: String,
-                                     repositoryOwnerName: String,
-                                     url: String,
-                                     version: SemanticVersion,
-                                     summary: String?) -> String {
-        createMessage(preamble: "⬆️ \(repositoryOwnerName) just released \(packageName) v\(version)",
-                      summary: summary,
-                      url: "\(url)#releases")
-    }
-
-    static func firehoseMessage(package: Joined<Package, Repository>,
-                                version: Version) -> String? {
-        guard let packageName = version.packageName,
-              let repoName = package.repository?.name,
-              let owner = package.repository?.owner,
-              let ownerName = package.repository?.ownerDisplayName,
-              let semVer = version.reference.semVer
-        else { return nil }
-        let url = SiteURL.package(.value(owner), .value(repoName), .none).absoluteURL()
-        return package.model.isNew
-        ? newPackageMessage(packageName: packageName,
-                            repositoryOwnerName: ownerName,
-                            url: url,
-                            summary: package.repository?.summary)
-        : versionUpdateMessage(packageName: packageName,
-                               repositoryOwnerName: ownerName,
-                               url: url,
-                               version: semVer,
-                               summary: package.repository?.summary)
-    }
-
-    static func postToFirehose(client: Client,
-                               package: Joined<Package, Repository>,
-                               version: Version) async throws {
-        guard Current.allowTwitterPosts() else {
-            throw Error.postingDisabled
-        }
-        guard let message = firehoseMessage(package: package, version: version) else {
-            throw Error.invalidMessage
-        }
-        try await Current.twitterPostTweet(client, message)
-    }
-
-    static func postToFirehose(client: Client,
-                               package: Joined<Package, Repository>,
-                               versions: [Version]) async throws {
-        let (release, preRelease, defaultBranch) = Package.findSignificantReleases(
-            versions: versions,
-            branch: package.repository?.defaultBranch
-        )
-        let idsLatest = [release, preRelease, defaultBranch].compactMap { $0?.id }
-        // filter on versions with a tag and which are in the "latest" triple
-        let versions = versions.filter { version in
-            guard version.reference.isTag,
-                  let id = version.id else { return false }
-            return idsLatest.contains(id)
-        }
-        var firstError: Swift.Error? = nil
-        for version in versions {
-            // Try all posts and record first error, if any
-            do {
-                try await postToFirehose(client: client, package: package, version: version)
-            } catch {
-                if firstError != nil {
-                    firstError = error
-                }
-            }
-        }
-        if let error = firstError {
-            throw error
+            throw Social.Error.requestFailed(response.status, response.body?.asString() ?? "")
         }
     }
